% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createCohorts}
\alias{createCohorts}
\title{Create and summarise the target and outcome cohorts}
\usage{
createCohorts(connectionDetails, cdmDatabaseschema, cohortDatabaseschema,
  cohortTable, targetId, outcomeId)
}
\arguments{
\item{cdmDatabaseschema}{The schema holding the CDM data}

\item{cohortDatabaseschema}{The schema holding the cohort table}

\item{cohortTable}{The name of the cohort table}

\item{targetId}{The cohort definition id of the target population}

\item{outcomeId}{The cohort definition id of the outcome}

\item{connectioDetails}{The connections details for connecting to the CDM}
}
\value{
A summary of the cohort counts
}
\description{
Create and summarise the target and outcome cohorts
}
\details{
This will create the risk prediciton cohorts and then count the table sizes
}
